feat: Base Mainnet deployment infrastructure ready
Created comprehensive deployment setup for Base Mainnet:

1. Deployment Scripts:
   - script/DeployBaseMainnet.s.sol - Solidity deployment with safety checks
   - deploy-mainnet-safe.sh - Interactive bash script with security validation

2. Security Features:
   - Compromised key detection (blocks exposed private key)
   - Balance verification (requires >= 0.05 ETH)
   - Network verification (confirms Base Mainnet Chain ID 8453)
   - 10-second countdown before deployment
   - Dry run option for testing

3. Documentation:
   - MAINNET_DEPLOYMENT_SECURITY.md - Complete security-first guide
   - DEPLOYMENT_READY.md - Quick start with checklist
   - BASE_MAINNET_ADDRESSES.md - Network addresses reference
   - MAINNET_DEPLOYMENT_GUIDE.md - Step-by-step instructions

4. Fixed Compilation Issues:
   - Removed Unicode characters from Solidity (emojis not supported)
   - Changed _getSettlementPrice from internal to public (for try/catch)
   - Fixed poolFee type casting in OptionsProtocolGasless
   - Installed Foundry dependencies (OpenZeppelin, Uniswap, forge-std)

5. Contract Status:
   - OptionsProtocol: 20,645 bytes (within 24,576 limit) ✅
   - Ready for mainnet deployment
   - Base Sepolia verified at 0xD7AFfB2B3303e9Cb44C9d9aFA6bD938200b3C8F2

CRITICAL SECURITY NOTE:
Private key 0xe9cc26bc179ca4303400a72a9d54ee1ff8af614a7811fc12ff5e235490b9ee4c
is COMPROMISED and blocked by deployment script. User must generate
new private key before mainnet deployment.

Ready to deploy with: ./deploy-mainnet-safe.sh

+++ src/OptionsProtocolGasless.sol
@@ -75,7 +75,7 @@ contract OptionsProtocolGasless is OptionsProtocol,IERC1271{
             uint256 poolFee = cfg.poolFee;
             address router = cfg.swapVenue;
             // Swap with 0 min out for now (front-end should compute safe min)
-            uint256 amountOut = _swapFromStablecoin(offer.underlying, requiredUSDC, offer.stablecoin, router, poolFee, 0);
+            uint256 amountOut = _swapFromStablecoin(offer.underlying, requiredUSDC, offer.stablecoin, router, uint24(poolFee), 0);
             collateralLocked = amountOut;
         } else {
             // For PUTs: keep USDC as collateral
